// src/backend/sfdc/model/SBQQ__Product__c.ts
export class SBQQ__Product__c {
  Id: string;
  Name: string;
  Description: string;
  ProductCode: string;
  SBQQ__Taxable__c: boolean;
  Apply_Houston_Franchise_Fee__c: boolean;
  Apply_Fuel_Surcharge__c: boolean;
  ApplyEnhancedSafetyFee__c: any;
  Apply_EEC__c: boolean;
  Asset_Summary__c: string;
  Number_of_Services__c?: number;
}